Move code to subcmd modules & unify exception handling
Fixes #136
