+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
   -  id: black
      args: ['--target-version', 'py38']
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.0.280
+  rev: v0.0.282
   hooks:
   - id: ruff
     args: [--fix, --exit-non-zero-on-fix]
