+++ src/voc4cat/transform.py
@@ -1,14 +1,25 @@
 import logging
 import shutil
+from collections import defaultdict
+from itertools import count, zip_longest
 from pathlib import Path
 from urllib.parse import urlsplit
 
+import openpyxl
+from openpyxl.styles import Alignment
 from rdflib import Graph
 from rdflib.namespace import SKOS
 
+from voc4cat import config
 from voc4cat.checks import Voc4catError
-from voc4cat.utils import EXCEL_FILE_ENDINGS, RDF_FILE_ENDINGS
-from voc4cat.wrapper import hierarchy_from_indent, hierarchy_to_indent, make_ids
+from voc4cat.util import (
+    dag_from_indented_text,
+    dag_from_narrower,
+    dag_to_narrower,
+    dag_to_node_levels,
+    get_concept_and_level_from_indented_line,
+)
+from voc4cat.utils import EXCEL_FILE_ENDINGS, RDF_FILE_ENDINGS, is_supported_template
 
 logger = logging.getLogger(__name__)
 
@@ -63,6 +74,272 @@ def read_split_turtle(vocab_dir: Path) -> Graph:
         cs_graph += graph
     cs_graph.serialize(destination=vocab_dir.with_suffix(".ttl"), format="turtle")
     return cs_graph
+
+
+def make_ids(
+    fpath: Path, outfile: Path, search_prefix: str, start_id: int, base_iri=None
+):
+    """
+    Replace all prefix:suffix CURIEs using IDs counting up from start_id
+
+    If base_iri is None the new IRI is a concatenation of VOC_BASE_IRI and ID.
+    If base_iri is given the new IRI is the concatenation of base_iri and ID.
+    """
+    logger.info("\nReplacing '%s' IRIs.", search_prefix)
+    # Load in data_only mode to get cell values not formulas.
+    wb = openpyxl.load_workbook(fpath, data_only=True)
+    is_supported_template(wb)
+    if base_iri is None:
+        base_iri = wb["Concept Scheme"].cell(row=2, column=2).value
+        if base_iri is None:
+            base_iri = "https://example.org/"
+            wb["Concept Scheme"].cell(row=2, column=2).value = base_iri
+            logger.warning("No concept scheme IRI found, using https://example.org/")
+
+    # TODO if config is set: Check that file name is in config
+    voc_config = config.IDRANGES.vocabs.get(fpath.stem, {})
+    id_length = voc_config.get("id_length", 7)
+
+    id_gen = count(start_id)
+    replaced_iris = {}
+    # process primary iri column in both worksheets
+    for sheet in ["Concepts", "Collections"]:
+        ws = wb[sheet]
+        # iterate over first two columns; skip header and start from row 3
+        for row in ws.iter_rows(min_row=3, max_col=2):  # pragma: no branch
+            if row[0].value:
+                iri = str(row[0].value)
+                if not iri.startswith(search_prefix):
+                    continue
+                if iri in replaced_iris:
+                    iri_new = replaced_iris[iri]
+                else:
+                    iri_new = base_iri + f"{next(id_gen):0{id_length}d}"
+                    msg = f"[{sheet}] Replaced CURIE {iri} by {iri_new}"
+                    logger.debug(msg)
+                    replaced_iris[iri] = iri_new
+                row[0].value = iri_new
+
+    # replace iri by new_iri in all columns where it might be referenced
+    cols_to_update = {
+        "Concepts": [6],
+        "Collections": [3],
+        "Additional Concept Features": [0, 1, 2, 3, 4, 5],
+    }
+    for sheet, cols in cols_to_update.items():
+        ws = wb[sheet]
+        # iterate over first two columns; skip header and start from row 3
+        for row in ws.iter_rows(min_row=3, max_col=1 + max(cols)):  # pragma: no branch
+            # stop processing a sheet after 3 empty rows
+            for col in cols:
+                if row[col].value:
+                    iris = [iri.strip() for iri in str(row[col].value).split(",")]
+                    count_new_iris = [1 for iri in iris if iri in replaced_iris].count(
+                        1
+                    )
+                    if count_new_iris:
+                        new_iris = [replaced_iris.get(iri, iri) for iri in iris]
+                        msg = (
+                            f"[{sheet}] Replaced {count_new_iris} CURIs "
+                            f"in cell {row[col].coordinate}"
+                        )
+                        logger.debug(msg)
+                        row[col].value = ", ".join(new_iris)
+                    # subsequent_empty_rows = 0
+
+    wb.save(outfile)
+    logger.info("Saved updated file as %s", outfile)
+    return 0
+
+
+def hierarchy_from_indent(fpath, outfile, sep):
+    """
+    Convert indentation hierarchy of concepts to children-URI form.
+
+    If separator character(s) are given they will be replaced with
+    the xlsx default indent. This is also the default if sep is None.
+    """
+    logger.info("Reading concepts from file %s", fpath)
+    # Load in data_only mode to get cell values not formulas.
+    wb = openpyxl.load_workbook(fpath, data_only=True)
+    is_supported_template(wb)
+    # process both worksheets
+    subsequent_empty_rows = 0
+    ws = wb["Concepts"]
+    concepts_indented = []
+    row_by_iri = defaultdict(dict)
+    # read concepts, determine their indentation level, then clear indentation
+    col_last = 9
+    max_row = 0
+    for row in ws.iter_rows(min_row=3, max_col=col_last):  # pragma: no branch
+        iri = row[0].value
+        if iri and row[1].value:
+            row_no = row[0].row
+            lang = row[2].value
+
+            if sep is None:  # xlsx indentation
+                level = int(row[1].alignment.indent)
+                ws.cell(row_no, column=2).alignment = Alignment(indent=0)
+            else:
+                descr, level = get_concept_and_level_from_indented_line(
+                    row[1].value, sep=sep
+                )
+                ws.cell(row_no, column=2).value = descr
+                ws.cell(row_no, column=2).alignment = Alignment(indent=0)
+            concepts_indented.append(level * " " + iri)
+
+            if iri in row_by_iri:  # merge needed
+                # compare fields, merge if one is empty, error if different values
+                new_data = [
+                    ws.cell(row_no, col_no).value for col_no in range(2, col_last)
+                ]
+                old_data = row_by_iri[iri].get(lang, [None] * (len(new_data)))
+                merged = []
+                for old, new in zip(old_data, new_data, strict=True):
+                    if (old and new) and (old != new):
+                        msg = f"Cannot merge rows for {iri}. Resolve differences manually."
+                        raise Voc4catError(msg)
+                    merged.append(old if old else new)
+                row_by_iri[iri][lang] = merged
+            else:
+                row_by_iri[iri][lang] = [
+                    ws.cell(row_no, col_no).value for col_no in range(2, col_last)
+                ]
+            max_row = row_no
+
+        # stop processing a sheet after 3 empty rows
+        elif subsequent_empty_rows < 2:  # noqa: PLR2004
+            subsequent_empty_rows += 1
+        else:
+            subsequent_empty_rows = 0
+            break
+
+    term_dag = dag_from_indented_text("\n".join(concepts_indented))
+    children_by_iri = dag_to_narrower(term_dag)
+
+    # Write rows to xlsx-file, write translation directly after each concept.
+    col_children_iri = 7
+    row = 3
+    for iri in children_by_iri:
+        for lang in row_by_iri[iri]:
+            ws.cell(row, column=1).value = iri
+            for col, stored_value in zip_longest(
+                range(2, col_last), row_by_iri[iri][lang]
+            ):
+                ws.cell(row, column=col).value = stored_value
+            ws.cell(row, column=col_children_iri).value = ", ".join(
+                children_by_iri[iri]
+            )
+            row += 1
+
+    while row <= max_row:
+        # Clear remaining rows.
+        for col in range(1, col_last):
+            ws.cell(row, column=col).value = None
+        row += 1
+
+    wb.save(outfile)
+    logger.info("Saved file with children-IRI hierarchy as %s", outfile)
+    return 0
+
+
+def hierarchy_to_indent(fpath, outfile, sep):
+    """
+    Convert concept hierarchy in children-URI form to indentation.
+
+    If separator character(s) are given they will be used.
+    If sep is None, xlsx default indent will be used.
+    """
+    logger.info("Reading concepts from file %s", fpath)
+    wb = openpyxl.load_workbook(fpath)
+    is_supported_template(wb)
+    ws = wb["Concepts"]
+
+    concept_children_dict = {}
+    subsequent_empty_rows = 0
+    row_by_iri = defaultdict(dict)
+    col_last = 9
+    # read all IRI, preferred labels, children_uris from the sheet
+    for rows_total, row in enumerate(  # pragma: no branch
+        ws.iter_rows(min_row=3, max_col=col_last, values_only=True)
+    ):
+        if row[0] and row[1]:
+            iri = row[0]
+            lang = row[2]
+            children_uris = [] if not row[6] else [c.strip() for c in row[6].split(",")]
+            # We need to check if ChildrenIRI, Provenance & Source Vocab URL
+            # are consistent across languages since SKOS has no support for
+            # per language statements. (SKOS-XL would add this)
+            if iri in row_by_iri:  # merge needed
+                # compare fields, merge if one is empty, error if different values
+                new_data = [row[col_no] for col_no in range(6, col_last)]
+                old_data = row_by_iri[iri][next(iter(row_by_iri[iri].keys()))][5:]
+                merged = []
+                for old, new in zip(old_data, new_data):
+                    if (old and new) and (old != new):
+                        msg = f'Merge conflict for concept {iri}. New: "{new}" - Before: "{old}"'
+                        raise Voc4catError(msg)
+                    merged.append(old if old else new)
+                row_by_iri[iri][lang] = [row[col] for col in range(1, 6)] + merged
+            else:
+                row_by_iri[iri][lang] = [row[col] for col in range(1, col_last)]
+                concept_children_dict[iri] = children_uris
+
+        # stop processing a sheet after 3 empty rows
+        elif subsequent_empty_rows < 2:  # noqa: PLR2004
+            subsequent_empty_rows += 1
+        else:
+            subsequent_empty_rows = 0
+            rows_total -= 2  # noqa: PLW2901
+            break
+    else:
+        pass
+
+    term_dag = dag_from_narrower(concept_children_dict)
+    concept_levels = dag_to_node_levels(term_dag)
+
+    # Write indented rows to xlsx-file but
+    # 1 - each concept translation only once
+    # 2 - details only once per concept and language
+    iri_written = []
+    row = 3
+    for iri, level in concept_levels:
+        for transl_no, lang in enumerate(row_by_iri[iri]):
+            if transl_no and (iri, lang) in iri_written:  # case 1
+                continue
+            ws.cell(row, 1).value = iri
+            concept_text = row_by_iri[iri][lang][0]
+            if sep is None:
+                ws.cell(row, 2).value = concept_text
+                ws.cell(row, 2).alignment = Alignment(indent=level)
+            else:
+                ws.cell(row, 2).value = sep * level + concept_text
+                ws.cell(row, 2).alignment = Alignment(indent=0)
+            ws.cell(row, 3).value = lang
+
+            if (iri, lang) in iri_written:  # case 2
+                for col, _ in zip_longest(
+                    range(4, col_last + 1), row_by_iri[iri][lang][2:]
+                ):
+                    ws.cell(row, column=col).value = None
+                row += 1
+                continue
+
+            for col, stored_value in zip_longest(
+                range(4, col_last + 1), row_by_iri[iri][lang][2:]
+            ):
+                ws.cell(row, column=col).value = stored_value
+            # clear children IRI column G
+            ws.cell(row, column=7).value = None
+            row += 1
+            iri_written.append((iri, lang))
+
+    wb.save(outfile)
+    logger.info("Saved file with indentation hierarchy as %s", outfile)
+    return 0
+
+
+# ===== transform command implementation & helpers to validate cmd options =====
 
 
 def _check_make_ids_args(args):
