+++ tests/test_util.py
@@ -1,4 +1,5 @@
 import pytest
+from voc4cat.checks import Voc4catError
 from voc4cat.util import (
     dag_from_indented_text,
     dag_from_narrower,
@@ -230,7 +231,7 @@ a2
 def test_undefined_child():
     narower = {"a1": [], "a2": ["c"]}
     with pytest.raises(
-        ValueError, match='Concept "c" needs to defined if used as narrower concept.'
+        Voc4catError, match='Concept "c" needs to defined if used as narrower concept.'
     ):
         dag_from_narrower(narower)
 
@@ -258,7 +259,7 @@ def test_none_as_sep():
 def test_bad_dedent():
     text = " x1\nx2"
     with pytest.raises(
-        ValueError, match='First line "x1" must be at lowest indentation level.'
+        Voc4catError, match='First line "x1" must be at lowest indentation level.'
     ):
         dag_from_indented_text(text)
 
@@ -266,7 +267,7 @@ def test_bad_dedent():
 def test_bad_indent():
     text = "x1\n--x2"
     with pytest.raises(
-        ValueError, match='Indentation increases by more than one level for "x2"'
+        Voc4catError, match='Indentation increases by more than one level for "x2"'
     ):
         dag_from_indented_text(text, sep="-")
